+++ lib/generators/ckeditor/templates/mongoid/carrierwave/ckeditor/asset.rb
@@ -1,7 +1,7 @@
 class Ckeditor::Asset
   include Ckeditor::Orm::Mongoid::AssetBase
 
-  delegate :url, :current_path, :size, :content_type, :filename, to: :data
+  delegate :url, :current_path, :content_type, to: :data
 
   validates :data, presence: true
 end