+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unrealized
 
   * Remove parameterize_filenames from paperclip backend (fix issue #738)
+  * Don't delegate size method to data, read it from cached attribute data_file_size
 
 ## Ckeditor 4.2.3