+++ CHANGES.md
@@ -1,10 +1,11 @@
 # Ckeditor changelog
 
-## Unrealized
+## Ckeditor 4.2.4
 
   * Remove parameterize_filenames from paperclip backend (fix issue #738)
   * Don't delegate size method to data, read it from cached attribute data_file_size
   * Update CKEditor to version 4.7.1 full (28 Jun 2017)
+  * Move method "assets_pipeline_enabled?" to main module and allow to change it from ckeditor config
 
 ## Ckeditor 4.2.3